EmpresaEJB actualizado

+++ Versiones/proyecto/trazabilidad-ejb/src/main/java/ejb/EmpresaEJB.java
@@ -28,14 +28,25 @@ public class EmpresaEJB implements GestionEmpresa {
 	private EntityManager em;
 
 	@Override
-	public void insertarEmpresa(Empresa empresa) throws ClienteExistenteException {
-		Empresa empresaExistente = em.find(Empresa.class, empresa.getID());
-		//Hace falta ver admin
-		if (empresaExistente != null) {
-			throw new ClienteExistenteException();
+	public void insertarEmpresa(UserApk user, Empresa empresa) throws ProyectoException {
+		
+		UserApk userExistente = em.find(UserApk.class, user.getUser());
+		
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
 		}
 		
-		em.persist(empresa);
+		if(user.isAdministrativo()) {
+			Empresa empresaExistente = em.find(Empresa.class, empresa.getID());
+
+			if (empresaExistente != null) {
+				throw new ClienteExistenteException();
+			}
+			
+			em.persist(empresa);
+		} else {
+			throw new UserNoAdminException();
+		}
 	}
 
 	@Override
@@ -45,37 +56,47 @@ public class EmpresaEJB implements GestionEmpresa {
 	}
 
 	@Override
-	public void actualizarEmpresa(Empresa empresa) throws ProyectoException {
-		Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
-		if (empresaEntity == null) {
-			throw new ClienteNoEncontradoException();
+	public void actualizarEmpresa(UserApk user, Empresa empresa) throws ProyectoException {
+		UserApk userExistente = em.find(UserApk.class, user.getUser());
+		
+		if (userExistente == null) {
+			throw new UserNoEncontradoException();
 		}
 		
-		empresaEntity.setCiudad(empresa.getCiudad());
-		empresaEntity.setCodPostal(empresa.getCodPostal());
-		empresaEntity.setCuentas(empresa.getCuentas());
-		empresaEntity.setFecha_Alta(empresa.getFecha_Alta());
-		empresaEntity.setDireccion(empresa.getDireccion());
-		empresaEntity.setIdent(empresa.getIdent());
-		empresaEntity.setPais(empresa.getPais());
-		empresaEntity.setTipo_cliente(empresa.getTipo_cliente());
-		empresaEntity.setFecha_Baja(empresa.getFecha_Baja());
-		
-		empresaEntity.setRazonSocial(empresa.getRazonSocial());
-		
-		TypedQuery<Segregada> query = em.createQuery("SELECT s FROM Segregada s", Segregada.class);
-		
-		for(Segregada s : query.getResultList()) {
-			for(CuentaFintech c : empresa.getCuentas()) {
-				if(s.getIBAN() == c.getIBAN()) {
-					throw new CuentaSegregadaYaAsignadaException();
-				} else {
-					empresaEntity.setCuentas(empresa.getCuentas());
+		if(user.isAdministrativo()) {
+			Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
+			if (empresaEntity == null) {
+				throw new ClienteNoEncontradoException();
+			}
+			
+			empresaEntity.setCiudad(empresa.getCiudad());
+			empresaEntity.setCodPostal(empresa.getCodPostal());
+			empresaEntity.setCuentas(empresa.getCuentas());
+			empresaEntity.setFecha_Alta(empresa.getFecha_Alta());
+			empresaEntity.setDireccion(empresa.getDireccion());
+			empresaEntity.setIdent(empresa.getIdent());
+			empresaEntity.setPais(empresa.getPais());
+			empresaEntity.setTipo_cliente(empresa.getTipo_cliente());
+			empresaEntity.setFecha_Baja(empresa.getFecha_Baja());
+			
+			empresaEntity.setRazonSocial(empresa.getRazonSocial());
+			
+			TypedQuery<Segregada> query = em.createQuery("SELECT s FROM Segregada s", Segregada.class);
+			
+			for(Segregada s : query.getResultList()) {
+				for(CuentaFintech c : empresa.getCuentas()) {
+					if(s.getIBAN() == c.getIBAN()) {
+						throw new CuentaSegregadaYaAsignadaException();
+					} else {
+						empresaEntity.setCuentas(empresa.getCuentas());
+					}
 				}
 			}
+			
+			em.merge(empresaEntity);
+		} else {
+			throw new UserNoAdminException();
 		}
-		
-		em.merge(empresaEntity);
 	}
 
 	@Override
@@ -99,56 +120,84 @@ public class EmpresaEJB implements GestionEmpresa {
 	}
 	
 	@Override
-	public void cerrarCuentaEmpresa(Empresa empresa) throws ProyectoException {
-		Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
-		
-		if (empresaEntity == null) {
-			throw new ClienteNoEncontradoException();
+	public void cerrarCuentaEmpresa(UserApk user, Empresa empresa) throws ProyectoException {
+
+		UserApk userApkEntity = em.find(UserApk.class, user.getUser());
+		if (userApkEntity == null) {
+			throw new UserNoEncontradoException();
 		}
 		
-		List<CuentaFintech> cuentasEmpresa = empresaEntity.getCuentas();
-		boolean ok = false;
-		
-		for(CuentaFintech c : cuentasEmpresa) {
-			if(c.getEstado() == true) {
-				ok = true;
+		if(user.isAdministrativo()) {
+			Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
+			
+			if (empresaEntity == null) {
+				throw new ClienteNoEncontradoException();
 			}
-		}
-		
-		if(!ok) {
-			empresaEntity.setEstado(ok);
+			
+			List<CuentaFintech> cuentasEmpresa = empresaEntity.getCuentas();
+			boolean ok = false;
+			
+			for(CuentaFintech c : cuentasEmpresa) {
+				if(c.getEstado() == true) {
+					ok = true;
+				}
+			}
+			
+			if(!ok) {
+				empresaEntity.setEstado(ok);
+			} else {
+				throw new NoBajaClienteException();
+			}
+			
+			em.merge(empresaEntity);	
 		} else {
-			throw new NoBajaClienteException();
+			throw new UserNoAdminException();
 		}
-		
-		em.merge(empresaEntity);
 	}
 
 	
 	@Override
-	public void eliminarEmpresa(Empresa empresa) throws ProyectoException {
-		Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
-		Cliente clienteEntity = em.find(Cliente.class, empresa.getID());
+	public void eliminarEmpresa(UserApk user, Empresa empresa) throws ProyectoException {
 		
-		if ((empresaEntity == null) && (clienteEntity == null)) {
-			throw new ClienteNoEncontradoException();
+		UserApk userApkEntity = em.find(UserApk.class, user.getUser());
+		if (userApkEntity == null) {
+			throw new UserNoEncontradoException();
 		}
 		
-		em.remove(empresaEntity);
-		em.remove(clienteEntity);
+		if(user.isAdministrativo()) {
+			Empresa empresaEntity = em.find(Empresa.class, empresa.getID());
+			
+			if (empresaEntity == null) {
+				throw new ClienteNoEncontradoException();
+			}
+			
+			em.remove(empresaEntity);	
+		} else {
+			throw new UserNoAdminException();
+		}
 	}
 
 	@Override
-	public void eliminarTodasEmpresas() throws ProyectoException {
-		List<Empresa> empresas = obtenerEmpresas();
+	public void eliminarTodasEmpresas(UserApk user) throws ProyectoException {
 		
-		for (Empresa e : empresas) {
-			Cliente clienteEntity = em.find(Cliente.class, e.getID());
-			em.remove(clienteEntity);
+		UserApk userApkEntity = em.find(UserApk.class, user.getUser());
+		if (userApkEntity == null) {
+			throw new UserNoEncontradoException();
 		}
 		
-		for (Empresa e : empresas) {
-			em.remove(e);
+		if(user.isAdministrativo()) {
+			List<Empresa> empresas = obtenerEmpresas();
+			
+			for (Empresa e : empresas) {
+				Cliente clienteEntity = em.find(Cliente.class, e.getID());
+				em.remove(clienteEntity);
+			}
+			
+			for (Empresa e : empresas) {
+				em.remove(e);
+			}
+		} else {
+			throw new UserNoAdminException();
 		}
 	}
